Enable proper 64-bit support in material group handling
The m_RemapData field in CKMaterialGroup now uses CKUINTPTR to safely
hold pointer values on both 32-bit and 64-bit architectures. This
prevents silent data loss when storing buffer pointers.

Additional type conversions have been added in the DX9 rasterizer,
configuration loader, and vertex cache optimizer to satisfy strict
compiler warnings about implicit narrowing conversions.

+++ src/CKRasterizer/CKDX9Rasterizer/CKDX9RasterizerContext.cpp
@@ -4883,10 +4883,10 @@ CKBOOL CKDX9RasterizerContext::CreateVertexDeclaration(CKDWORD VFormat, LPDIRECT
         declaration[elementIndex].Type = declType;
         declaration[elementIndex].Method = D3DDECLMETHOD_DEFAULT;
         declaration[elementIndex].Usage = usage;
-        declaration[elementIndex].UsageIndex = i;
+        declaration[elementIndex].UsageIndex = (BYTE)i;
 
         // Update offset for next element
-        offset += texDimension * sizeof(float);
+        offset += (WORD)(texDimension * sizeof(float));
         elementIndex++;
     }
